Multilingual-AI-Assistant-main/requirements.txt
SpeechRecognition
pipwin
pyaudio
gTTS
google-generativeai
python-dotenv
streamlit
PyPDF2
langdetect
flask
flask-cors
pdf2image
pytesseract
Pillow
backoff

-e .